core: silence MSVC compiler warning
The return type of wcore_attrib_list_length() is int32_t and at the same
time MSVC promotes the (int32_t) math at the return line to int64_t thus
causing a warning message. Explicitly cast it to make MSVC happy.

+++ src/waffle/core/wcore_attrib_list.c
@@ -40,7 +40,7 @@ wcore_attrib_list_length(const int32_t attrib_list[])
     while (*i != 0)
         i += 2;
 
-    return (i - attrib_list) / 2;
+    return (int32_t) (i - attrib_list) / 2;
 }
 
 bool
